fix: limit the code that is run in the thread pool
Prior to this change, things were run in the thread pool that could
also run on the main thread (as they were blocking the main thread
anyway).

With this change there is more control on which things should be run
on the thread pool. This way we can make sure that all calls on `wait()`
are outside the thread pool, as else they can lead to a deadlock.