committing failing trial w/ common trait Writable
This approach doesn't work as E::Fqk is only a Field and implementing the trait
over Fqk12 doesn't work. Implementing for Compress doesn't work either because
of un-constrained type (G1Affine might implement Compress as well).
The real fix would be to make all types in paired implement Compress (or at
least the relevant one) so there would be one common trait interface to
serialize/unserialize elements. Currently there are 3 different traits related
to this functionality which is disturbingly annoying.