update scheduler git branch
use GPUSelector, logging messages and makes scheduler an optional
dependency

makes the scheduler an optional dependency

New type used to indicate the type of task being executed in bellperson, this allows to hide the scheduler from upstream users of the library

removes the in_priority functions which does not make sense as users can use the TaskType enum.

update the prover and prover_ext so that the latter is use only if the gpu feature is used

update test so using winning_post type indicates a higher priority task

encode the number of gpus to use in the macro call.